gateware_1chain/cocotb/test_modulate.py
# test_modulate.py

import cocotb
from cocotb.triggers import RisingEdge, FallingEdge
from cocotb.clock import Clock

@cocotb.test()
async def test_modulate(dut):
    """ Testbench for Modulate module """

    # Create a clock on dut.clk with a period of 10ns
    cocotb.fork(Clock(dut.clk, 10, units="ns").start())

    # Reset the design
    dut.reset <= 1
    await RisingEdge(dut.clk)
    dut.reset <= 0

    # Add further checks here to verify the behavior of modulate.v
    # Example: assert dut.output_signal.value == expected_value
    for _ in range(100):
        await RisingEdge(dut.clk)
        # Monitor outputs and assert values
